Rename feature package to rule
